silence more 32-bit warnings

+++ src/includes/joedb/Readonly_Journal.h
@@ -12,10 +12,10 @@ namespace joedb
   public:
    Readonly_Journal(Generic_File &file);
 
-   uint64_t get_checkpoint_position() const {return checkpoint_position;}
+   size_t get_checkpoint_position() const {return checkpoint_position;}
    void replay_log(Writeable &writeable);
    void rewind();
-   void play_until(Writeable &writeable, uint64_t end);
+   void play_until(Writeable &writeable, size_t end);
    void play_until_checkpoint(Writeable &writeable)
    {
     play_until(writeable, checkpoint_position);
@@ -28,7 +28,7 @@ namespace joedb
   protected:
    Generic_File &file;
    unsigned checkpoint_index;
-   uint64_t checkpoint_position;
+   size_t checkpoint_position;
 
    Table_Id table_of_last_operation;
    Record_Id record_of_last_operation;
